Add support for Go develop containers

+++ docker/base/bootstrap_pure.sh
@@ -4,21 +4,26 @@
 # not only a few pre-built Go cross compilers, but rather bootstraps all of the
 # supported platforms from the origin Linux amd64 distribution.
 #
-# Usage: bootstrap.sh
+# Usage: bootstrap_pure.sh
 #
-# Needed environment variables:
+# Environment variables for remote bootstrapping:
 #   FETCH          - Remote file fetcher and checksum verifier (injected by image)
 #   ROOT_DIST      - 64 bit Linux Go binary distribution package
 #   ROOT_DIST_SHA1 - 64 bit Linux Go distribution package checksum
+#
+# Environment variables for local bootstrapping:
+#   GOROOT - Path to the lready installed Go runtime
 set -e
 
-# Download, verify and install the root distribution
-$FETCH $ROOT_DIST  $ROOT_DIST_SHA1
+# Download, verify and install the root distribution if pulled remotely
+if [ "$GOROOT" == "" ]; then
+  $FETCH $ROOT_DIST $ROOT_DIST_SHA1
 
-tar -C /usr/local -xzf `basename $ROOT_DIST`
-rm -f `basename $ROOT_DIST`
+  tar -C /usr/local -xzf `basename $ROOT_DIST`
+  rm -f `basename $ROOT_DIST`
 
-export GOROOT=/usr/local/go
+  export GOROOT=/usr/local/go
+fi
 export GOROOT_BOOTSTRAP=$GOROOT
 
 # Pre-build all guest distributions based on the root distribution
